feat: author, password 추가구현

+++ src/main/java/likelion/side_project_blog/service/ArticleService.java
@@ -1,25 +1,20 @@
 package likelion.side_project_blog.service;
 
-import jakarta.persistence.EntityNotFoundException;
 import likelion.side_project_blog.domain.Article;
 import likelion.side_project_blog.dto.request.AddArticleRequest;
+import likelion.side_project_blog.dto.request.DeleteRequest;
 import likelion.side_project_blog.dto.request.UpdateArticleRequest;
-import likelion.side_project_blog.dto.response.ApiResponse;
 import likelion.side_project_blog.dto.response.ArticleResponse;
 import likelion.side_project_blog.dto.response.CommentResponse;
 import likelion.side_project_blog.exception.ArticleNotFoundException;
+import likelion.side_project_blog.exception.PermissionDeniedException;
 import likelion.side_project_blog.repository.ArticleRepository;
 import likelion.side_project_blog.repository.CommentRepository;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -54,9 +49,12 @@ public class ArticleService {
 
 
     //글 삭제
-    public void deleteArticle(Long id){
+    public void deleteArticle(Long id, DeleteRequest request){
         Article article=articleRepository.findById(id)
                 .orElseThrow(()-> new ArticleNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
+        if(!article.getPassword().equals(request.getPassword())){
+            throw new PermissionDeniedException("해당 글에 대한 삭제 권한이 없습니다.");
+        }
         articleRepository.deleteById(id);
 
     }
@@ -65,6 +63,9 @@ public class ArticleService {
     public void updateArticle(Long id, UpdateArticleRequest request){
         Article article=articleRepository.findById(id)
                 .orElseThrow(()-> new ArticleNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
+        if(!article.getPassword().equals(request.getPassword())){
+            throw new PermissionDeniedException("해당 글에 대한 삭제 권한이 없습니다.");
+        }
         article.update(request.getTitle(),request.getContent());
         articleRepository.save(article);
